* Fixed compiler warnings about size_t using GCC 4.2.1.

+++ include/uhttp/http/HTTPServer.h
@@ -58,8 +58,8 @@ public:
 
   HTTP::StatusCode performRequestListener(HTTPRequest *httpReq) {
     HTTP::StatusCode lastStatusCode = HTTP::INTERNAL_SERVER_ERROR;
-    int listenerSize = httpRequestListenerList.size();
-    for (int n = 0; n < listenerSize; n++) {
+    size_t listenerSize = httpRequestListenerList.size();
+    for (size_t n = 0; n < listenerSize; n++) {
       HTTPRequestListener *listener = (HTTPRequestListener *)httpRequestListenerList.get(n);
       lastStatusCode = listener->httpRequestRecieved(httpReq);
       if (lastStatusCode == HTTP::OK_REQUEST)
